feat:refactor the code
